Rename db into datasore
